requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
blis==0.4.1
boto==2.49.0
boto3==1.10.46
botocore==1.13.46
bpemb==0.3.0
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
cycler==0.10.0
cymem==2.0.3
decorator==4.4.1
defusedxml==0.6.0
Deprecated==1.2.7
docutils==0.15.2
entrypoints==0.3
faiss-cpu==1.6.1
flair==0.4.3
future==0.18.2
gensim==3.8.1
hyperopt==0.2.2
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.2
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterlab==1.2.4
jupyterlab-server==1.0.6
kiwisolver==1.1.0
langdetect==1.0.7
langid==1.1.6
llvmlite==0.32.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.0.2
mpld3==0.3
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.2
nltk==3.4.5
notebook==6.0.2
numba==0.49.0
numpy==1.18.0
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
progress==1.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py==1.8.1
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.6
pytest==5.3.2
python-dateutil==2.8.1
pytorch-transformers==1.2.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.6.0
regex==2019.12.20
requests==2.22.0
s3transfer==0.2.1
sacremoses==0.0.35
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.9.0
segtok==1.5.7
Send2Trash==1.5.0
sentencepiece==0.1.85
six==1.13.0
sklearn==0.0
smart-open==1.9.0
spacy==2.2.4
sqlitedict==1.6.0
srsly==1.0.2
tabulate==0.8.6
tbb==2020.0.133
terminado==0.8.3
testpath==0.4.4
thinc==7.4.0
tiny-tokenizer==3.1.0
torch==1.3.1
tornado==6.0.3
tqdm==4.41.1
traitlets==4.3.3
umap-learn==0.4.2
urllib3==1.24.3
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
yml==0.0.1
zipp==0.6.0
